Small modifications, but big differences! (Read desc.)
Now, this physics works! The next step is to implement the collision circle of the outer wall, so the karts can't get away from it.

Also, it's necessary to discuss if these physics will be implemented on the final project or what, so we can decide to continue on the previous version, or on this one.