Added "a lot" of stuff, and optimization changes
Briefly, I optimized a little bit the game, 'cause it was lagging. Also, I enhanced the textures of the karts, but reduced its dimensions in order to optimize as well. I fixed some models of the karts, and re-organize some lines of code in order to keep it clean.

The next and final step is to add the skybox